refactor: 테스트에서 중복되는 부분 Fixture로 분리
