+++ src/test/java/woowacourse/fixture/ResponseFixture.java
@@ -1,0 +1,64 @@
+package woowacourse.fixture;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import java.util.List;
+import java.util.stream.Collectors;
+import org.springframework.http.HttpStatus;
+import woowacourse.shoppingcart.dto.response.CartItemResponse;
+import woowacourse.shoppingcart.dto.response.OrderResponse;
+
+public class ResponseFixture {
+
+    public static void 상품_추가됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
+    }
+
+    public static void 상품_삭제됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+
+    public static void 장바구니_아이템_추가됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
+    }
+
+    public static void 장바구니_아이템_목록_응답됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    public static void 장바구니_아이템_목록_포함됨(ExtractableResponse<Response> response, Long... productIds) {
+        List<Long> resultProductIds = response.jsonPath().getList(".", CartItemResponse.class).stream()
+                .map(CartItemResponse::getProductId)
+                .collect(Collectors.toList());
+        assertThat(resultProductIds).contains(productIds);
+    }
+
+    public static void 장바구니_삭제됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+
+    public static void 주문하기_성공함(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
+    }
+
+    public static void 주문_조회_응답됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    public static void 주문_내역_포함됨(ExtractableResponse<Response> response, Long... orderIds) {
+        List<Long> resultOrderIds = response.jsonPath().getList(".", OrderResponse.class).stream()
+                .map(OrderResponse::getId)
+                .collect(Collectors.toList());
+        assertThat(resultOrderIds).contains(orderIds);
+    }
+
+    public static void 주문_조회됨(ExtractableResponse<Response> response, Long orderId) {
+        OrderResponse resultOrder = response.as(OrderResponse.class);
+        assertThat(resultOrder.getId()).isEqualTo(orderId);
+    }
+}
