+++ src/test/java/woowacourse/shoppingcart/acceptance/CartAcceptanceTest.java
@@ -2,22 +2,24 @@ package woowacourse.shoppingcart.acceptance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static woowacourse.shoppingcart.acceptance.ProductAcceptanceTest.상품_등록되어_있음;
+import static woowacourse.fixture.RequestFixture.로그인_및_토큰_발급;
+import static woowacourse.fixture.RequestFixture.상품_등록되어_있음;
+import static woowacourse.fixture.RequestFixture.장바구니_삭제_요청;
+import static woowacourse.fixture.RequestFixture.장바구니_아이템_목록_조회_요청;
+import static woowacourse.fixture.RequestFixture.장바구니_아이템_수량_변경_요청;
+import static woowacourse.fixture.RequestFixture.장바구니_아이템_추가_요청;
+import static woowacourse.fixture.RequestFixture.장바구니_아이템_추가되어_있음;
+import static woowacourse.fixture.ResponseFixture.장바구니_삭제됨;
+import static woowacourse.fixture.ResponseFixture.장바구니_아이템_목록_응답됨;
+import static woowacourse.fixture.ResponseFixture.장바구니_아이템_목록_포함됨;
+import static woowacourse.fixture.ResponseFixture.장바구니_아이템_추가됨;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import woowacourse.auth.dto.LoginRequest;
-import woowacourse.auth.dto.LoginResponse;
 import woowacourse.shoppingcart.dto.response.CartItemResponse;
 
 @DisplayName("장바구니 관련 기능")
@@ -116,89 +118,5 @@ public class CartAcceptanceTest extends AcceptanceTest {
                 () -> assertThat(cartItemResponse.getProductId()).isEqualTo(productId1),
                 () -> assertThat(cartItemResponse.getQuantity()).isEqualTo(2)
         );
-    }
-
-    private String 로그인_및_토큰_발급(String name, String password) {
-        return RestAssured
-                .given().log().all()
-                .body(new LoginRequest(name, password))
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/api/login")
-                .then().log().all().extract()
-                .as(LoginResponse.class).getAccessToken();
-    }
-
-    public static ExtractableResponse<Response> 장바구니_아이템_추가_요청(String accessToken, Long productId) {
-        Map<String, Object> requestBody = new HashMap<>();
-        requestBody.put("id", productId);
-
-        return RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(requestBody)
-                .when().post("/api/customers/me/carts")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 장바구니_아이템_목록_조회_요청(String accessToken) {
-        return RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/api/customers/me/carts")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 장바구니_삭제_요청(String accessToken, Long cartId) {
-        return RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().delete("/api/customers/me/carts/{cartId}", cartId)
-                .then().log().all()
-                .extract();
-    }
-
-    private ExtractableResponse<Response> 장바구니_아이템_수량_변경_요청(String accessToken, Long cartId, int quantity) {
-        Map<String, Object> requestBody = new HashMap<>();
-        requestBody.put("quantity", quantity);
-
-        return RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(requestBody)
-                .when().patch("/api/customers/me/carts/" + cartId)
-                .then().log().all()
-                .extract();
-    }
-
-    public static void 장바구니_아이템_추가됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.header("Location")).isNotBlank();
-    }
-
-    public static Long 장바구니_아이템_추가되어_있음(String accessToken, Long productId) {
-        ExtractableResponse<Response> response = 장바구니_아이템_추가_요청(accessToken, productId);
-        return Long.parseLong(response.header("Location").split("/carts/")[1]);
-    }
-
-    public static void 장바구니_아이템_목록_응답됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 장바구니_아이템_목록_포함됨(ExtractableResponse<Response> response, Long... productIds) {
-        List<Long> resultProductIds = response.jsonPath().getList(".", CartItemResponse.class).stream()
-                .map(CartItemResponse::getProductId)
-                .collect(Collectors.toList());
-        assertThat(resultProductIds).contains(productIds);
-    }
-
-    public static void 장바구니_삭제됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
     }
 }
