+++ src/test/java/woowacourse/shoppingcart/acceptance/OrderAcceptanceTest.java
@@ -1,10 +1,17 @@
 package woowacourse.shoppingcart.acceptance;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static woowacourse.shoppingcart.acceptance.CartAcceptanceTest.장바구니_아이템_추가되어_있음;
-import static woowacourse.shoppingcart.acceptance.ProductAcceptanceTest.상품_등록되어_있음;
+import static woowacourse.fixture.RequestFixture.로그인_및_토큰_발급;
+import static woowacourse.fixture.RequestFixture.상품_등록되어_있음;
+import static woowacourse.fixture.RequestFixture.장바구니_아이템_추가되어_있음;
+import static woowacourse.fixture.RequestFixture.주문_내역_조회_요청;
+import static woowacourse.fixture.RequestFixture.주문_단일_조회_요청;
+import static woowacourse.fixture.RequestFixture.주문하기_요청;
+import static woowacourse.fixture.RequestFixture.주문하기_요청_성공되어_있음;
+import static woowacourse.fixture.ResponseFixture.주문_내역_포함됨;
+import static woowacourse.fixture.ResponseFixture.주문_조회_응답됨;
+import static woowacourse.fixture.ResponseFixture.주문_조회됨;
+import static woowacourse.fixture.ResponseFixture.주문하기_성공함;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.Arrays;
@@ -15,12 +22,7 @@ import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import woowacourse.auth.dto.LoginRequest;
-import woowacourse.auth.dto.LoginResponse;
 import woowacourse.shoppingcart.dto.request.CreateOrderDetailRequest;
-import woowacourse.shoppingcart.dto.response.OrderResponse;
 
 @DisplayName("주문 관련 기능")
 public class OrderAcceptanceTest extends AcceptanceTest {
@@ -92,74 +94,5 @@ public class OrderAcceptanceTest extends AcceptanceTest {
         // then
         주문_조회_응답됨(response);
         주문_조회됨(response, orderId);
-    }
-
-    private String 로그인_및_토큰_발급(String name, String password) {
-        return RestAssured
-                .given().log().all()
-                .body(new LoginRequest(name, password))
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/api/login")
-                .then().log().all().extract()
-                .as(LoginResponse.class).getAccessToken();
-    }
-
-    public static ExtractableResponse<Response> 주문하기_요청(List<CreateOrderDetailRequest> orderRequests,
-                                                        String accessToken) {
-        return RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(orderRequests)
-                .when().post("/api/customers/me/orders")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 주문_내역_조회_요청(String accessToken) {
-        return RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/api/customers/me/orders")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 주문_단일_조회_요청(Long orderId, String accessToken) {
-        return RestAssured
-                .given().log().all()
-                .auth().oauth2(accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/api/customers/me/orders/{orderId}", orderId)
-                .then().log().all()
-                .extract();
-    }
-
-    public static void 주문하기_성공함(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.header("Location")).isNotBlank();
-    }
-
-    public static Long 주문하기_요청_성공되어_있음(List<CreateOrderDetailRequest> orderRequests, String accessToken) {
-        ExtractableResponse<Response> response = 주문하기_요청(orderRequests, accessToken);
-        return Long.parseLong(response.header("Location").split("/orders/")[1]);
-    }
-
-    public static void 주문_조회_응답됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 주문_내역_포함됨(ExtractableResponse<Response> response, Long... orderIds) {
-        List<Long> resultOrderIds = response.jsonPath().getList(".", OrderResponse.class).stream()
-                .map(OrderResponse::getId)
-                .collect(Collectors.toList());
-        assertThat(resultOrderIds).contains(orderIds);
-    }
-
-    private void 주문_조회됨(ExtractableResponse<Response> response, Long orderId) {
-        OrderResponse resultOrder = response.as(OrderResponse.class);
-        assertThat(resultOrder.getId()).isEqualTo(orderId);
     }
 }
